+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added MOTD.
+
+### Changed
+
+- Changed level of dev console initialized message from info to debug.
+
 ## [1.1.0] - 2024-11-13
 
 ### Added
@@ -26,6 +34,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added LICENSE.
 - Added README.md.
 
-[unreleased]: https://github.com/Logtism/godotdevconsole/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/Logtism/godotdevconsole/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/Logtism/godotdevconsole/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/Logtism/godotdevconsole/releases/tag/v1.0.0