+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+## [1.2.0] - 2025-01-10
+
+### Added
+
 - Added MOTD.
 - Added Info command.
 - Added ActiveChanged event.
@@ -39,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added LICENSE.
 - Added README.md.
 
-[unreleased]: https://github.com/Logtism/godotdevconsole/compare/v1.1.0...HEAD
+[unreleased]: https://github.com/Logtism/godotdevconsole/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/Logtism/godotdevconsole/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/Logtism/godotdevconsole/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/Logtism/godotdevconsole/releases/tag/v1.0.0